+++ components/ssRewards/ssRewardsTable.js
@@ -402,7 +402,7 @@ const useStyles = makeStyles((theme) => {
       position: 'absolute',
       left: '0px',
       top: '0px',
-      border: '2px solid #DBE6EC',
+      outline: '2px solid #DBE6EC',
       background: '#13B5EC',
       borderRadius: '30px',
     },
@@ -411,7 +411,7 @@ const useStyles = makeStyles((theme) => {
       left: '28px',
       zIndex: '1',
       top: '0px',
-      border: '2px solid #DBE6EC',
+      outline: '2px solid #DBE6EC',
       background: '#13B5EC',
       borderRadius: '30px',
     },
