+++ components/ssSwap/ssSwap.js
@@ -1,6 +1,5 @@
 import Form from '../../ui/Form';
 import Setup from './setup';
-import classes from './ssSwap.module.css';
 
 function Swap() {
   return (
