+++ pages/liquidity/liquidity.module.css
@@ -25,6 +25,7 @@
 }
 
 .mainDescBg {
+  position: relative;
   display: flex;
   align-items: center;
   justify-content: center;
