+++ pages/migrate/migrate.module.css
@@ -869,6 +869,7 @@
 }
 
 .mainDescBg {
+  position: relative;
   display: flex;
   align-items: center;
   justify-content: center;
