+++ pages/swap/swap.module.css
@@ -81,6 +81,7 @@
 }
 
 .mainDescBg {
+  position: relative;
   display: flex;
   align-items: center;
   justify-content: center;
