+++ pages/vest/vest.module.css
@@ -85,6 +85,7 @@
 }
 
 .mainDescBg {
+  position: relative;
   display: flex;
   align-items: center;
   justify-content: center;
