+++ pages/whitelist/whitelist.module.css
@@ -37,6 +37,7 @@
 }
 
 .mainDescBg {
+  position: relative;
   display: flex;
   align-items: center;
   justify-content: center;
